refactored controller names
