+++ src/main/java/com/lgtm/easymoney/controllers/UserController.java
@@ -1,67 +1,93 @@
 package com.lgtm.easymoney.controllers;
 
-import com.lgtm.easymoney.models.Account;
 import com.lgtm.easymoney.models.User;
+import com.lgtm.easymoney.payload.BalanceReq;
+import com.lgtm.easymoney.payload.BalanceRsp;
 import com.lgtm.easymoney.payload.ErrorRsp;
-import com.lgtm.easymoney.payload.RegisterReq;
-import com.lgtm.easymoney.payload.RegisterRsp;
-import com.lgtm.easymoney.repositories.AccountRepository;
-import com.lgtm.easymoney.repositories.UserRepository;
+import com.lgtm.easymoney.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.Errors;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
+
 import javax.validation.Valid;
+import java.math.BigDecimal;
 import java.util.ArrayList;
-import java.util.List;
+import java.util.Arrays;
 
 @RestController
-@RequestMapping("/user")
+@RequestMapping("/balance")
 public class UserController {
-    private final UserRepository userRepository;
-    private final AccountRepository accountRepository;
-
     @Autowired
-    public UserController(UserRepository userRepository, AccountRepository accountRepository) {
-        this.userRepository = userRepository;
-        this.accountRepository = accountRepository;
+    private UserService userService;
+    @PutMapping("/deposit")
+    public ResponseEntity<?> deposit(@Valid @RequestBody BalanceReq req) {
+        // get params
+        Long uid = req.getUid();
+        BigDecimal amount = req.getAmount();
+        // make a deposit
+        User user = userService.getUserByID(uid);
+        boolean success = userService.makeADeposit(user, amount);
+        // payload
+        BalanceRsp res = new BalanceRsp();
+        res.setSuccess(success);
+        res.setCurrBalance(user.getBalance());
+
+        return success ? ResponseEntity.status(HttpStatus.OK).body(res) :
+                        ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ErrorRsp(new ArrayList<>(Arrays.asList("i think we need to refactor this")), "Error in making a deposit."));
+
+    }
+    @PutMapping("/withdraw")
+    public ResponseEntity<?> withdraw(@Valid @RequestBody BalanceReq req) {
+        // get params
+        Long uid = req.getUid();
+        BigDecimal amount = req.getAmount();
+        // make a withdraw
+        User user = userService.getUserByID(uid);
+        boolean success = userService.makeAWithdraw(user, amount);
+        // payload
+        BalanceRsp res = new BalanceRsp();
+        res.setSuccess(success);
+        res.setCurrBalance(user.getBalance());
+
+        return success ? ResponseEntity.status(HttpStatus.OK).body(res) :
+                ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ErrorRsp(new ArrayList<>(Arrays.asList("i think we need to refactor this")), "Error in making a withdraw."));
+
     }
 
-    @PostMapping("/register")
-    public ResponseEntity<?> register(@Valid @RequestBody RegisterReq registerReq, Errors errors) {
-        if (userRepository.existsByEmail(registerReq.getEmail())) {
-            List<String> errorFields = new ArrayList<>();
-            errorFields.add("email");
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new ErrorRsp(errorFields, "Email already registered!"));
-        }
-        if (accountRepository.existsByNumberAndRoutingNumber(registerReq.getAccountNumber(), registerReq.getRoutingNumber())) {
-            List<String> errorFields = new ArrayList<>();
-            errorFields.add("accountNumber");
-            errorFields.add("rountingNumber");
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new ErrorRsp(errorFields, "Bank account already registered!"));
-        }
-
-        var user = new User();
-        user.setEmail(registerReq.getEmail());
-        user.setPassword(registerReq.getPassword());  // TODO spring security password encoding
-        user.setTypeByStr(registerReq.getUserType());
-        user.setPhone(registerReq.getPhone());
-        user.setAddress(registerReq.getAddress());
-
-        var account = new Account();
-        account.setName(registerReq.getAccountName());
-        account.setNumber(registerReq.getAccountNumber());
-        account.setRoutingNumber(registerReq.getRoutingNumber());
-        user.setAccount(account);
-
-        var userCreated = userRepository.save(user);
-
-        return ResponseEntity.status(HttpStatus.CREATED).body(new RegisterRsp(userCreated.getId()));
-    }
-
-    // TODO login, get profile, update profile
+//    @PutMapping("/update")
+//    public ResponseEntity<?> update(@Valid @RequestBody BalanceReq balanceReq) {
+//        var userWrapper = userRepository.findById(balanceReq.getUid());
+//        if (userWrapper.isEmpty()) {
+//            List<String> errorFields = new ArrayList<>();
+//            errorFields.add("uid");
+//            // TODO load the current authenticated user, no need to check existence
+//            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ErrorRsp(errorFields, "User not found!"));
+//        }
+//
+//        // TODO maybe better to move the logics below to service layer?
+//        var user = userWrapper.get();
+//        var balance = user.getBalance();
+//        var amount = balanceReq.getAmount();
+//        var rsp = new BalanceRsp();
+//
+//        if (balanceReq.getIsDeposit()) {
+//            balance = balance.add(amount);
+//            user.setBalance(balance);
+//            rsp.setSuccess(Boolean.TRUE);
+//        } else {
+//            if (balance.compareTo(amount) < 0) {
+//                rsp.setSuccess(Boolean.FALSE);
+//            } else {
+//                balance = balance.subtract(amount);
+//                user.setBalance(balance);
+//                rsp.setSuccess(Boolean.TRUE);
+//            }
+//        }
+//
+//        userRepository.save(user);
+//        rsp.setCurrBalance(balance);
+//
+//        return ResponseEntity.status(HttpStatus.OK).body(rsp);
+//    }
 }
